requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
boto3==1.34.16
botocore==1.34.16
certifi==2023.11.17
cftime==1.6.3
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
cmake==3.25.0
comm==0.2.1
contourpy==1.1.1
copulas==0.9.2
cramjam==2.8.0
ctgan==0.8.0
cycler==0.12.1
dask==2023.5.0
dask-geopandas==0.3.1
debugpy==1.8.0
decorator==5.1.1
deepecho==0.5.0
distributed==2023.5.0
executing==2.0.1
Faker==19.13.0
fastjsonschema==2.19.1
fastparquet==2023.10.1
filelock==3.13.1
fiona==1.9.5
fonttools==4.47.2
fsspec==2023.12.2
geopandas==0.13.2
graphviz==0.20.1
h5netcdf==1.1.0
h5py==3.10.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kmeans-gpu==0.0.5
kmeans-pytorch==0.3
lit==15.0.7
llvmlite==0.41.1
locket==1.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
nbformat==5.9.2
nest-asyncio==1.5.8
netCDF4==1.6.5
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.0.3
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rdt==1.9.1
referencing==0.32.1
requests==2.28.1
rpds-py==0.17.1
s3transfer==0.10.0
scikit-learn==1.3.2
scipy==1.10.1
sdmetrics==0.13.0
sdv==1.9.0
seaborn==0.13.1
shapely==2.0.2
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.12
tblib==3.0.0
tenacity==8.2.3
threadpoolctl==3.2.0
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.0.0
tslearn==0.6.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
wcwidth==0.2.13
xarray==2023.1.0
zict==3.0.0
zipp==3.17.0
